move main module into app module
