+++ app/build.gradle
@@ -56,7 +56,7 @@ android {
             proguardFiles getDefaultProguardFile("proguard-android.txt"), "proguard-rules.pro"
         }
     }
-    dynamicFeatures = [":station", ":main"]
+    dynamicFeatures = [":station"]
 }
 
 dependencies {
