+++ app/src/main/AndroidManifest.xml
@@ -16,6 +16,32 @@
             android:roundIcon="@mipmap/ic_launcher_round"
             android:supportsRtl="true"
             android:theme="@style/AppTheme">
+        <activity
+                android:name="com.mxmariner.main.activity.MainActivity"
+                android:theme="@style/AppTheme">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+            <intent-filter
+                    android:autoVerify="true"
+                    android:order="1">
+                <action android:name="android.intent.action.VIEW" />
+                <category android:name="android.intent.category.BROWSABLE" />
+                <category android:name="android.intent.category.DEFAULT" />
+                <data android:scheme="http" />
+                <data android:scheme="https" />
+                <data android:host="mxmariner.com" />
+                <data android:path="/tides" />
+                <data android:pathPattern="/tides?.*" />
+            </intent-filter>
+            <meta-data
+                    android:name="default-url"
+                    android:value="https://mxmariner.com/tides" />
+        </activity>
+        <activity
+                android:name="com.mxmariner.main.activity.LocationSearchActivity"
+                android:theme="@style/TransparentTheme" />
     </application>
 
 </manifest>
