+++ app/src/main/java/com/mxmariner/main/model/TidesViewState.kt
@@ -1,4 +1,4 @@
-package com.mxmariner.tides.main.model
+package com.mxmariner.main.model
 
 sealed class TidesViewState
 data class TidesViewStateLoadingStarted(val message: String? = null) : TidesViewState()
