+++ app/src/main/java/com/mxmariner/main/viewmodel/TidesViewModel.kt
@@ -1,4 +1,4 @@
-package com.mxmariner.tides.main.viewmodel
+package com.mxmariner.main.viewmodel
 
 import android.content.res.Resources
 import androidx.lifecycle.ViewModel
@@ -7,10 +7,10 @@ import com.github.salomonbrys.kodein.Kodein
 import com.github.salomonbrys.kodein.instance
 import com.mxmariner.mxtide.api.StationType
 import com.mxmariner.tides.R
-import com.mxmariner.tides.main.adapter.TidesRecyclerAdapter
-import com.mxmariner.tides.main.model.TidesViewState
-import com.mxmariner.tides.main.model.TidesViewStateLoadingComplete
-import com.mxmariner.tides.main.model.TidesViewStateLoadingStarted
+import com.mxmariner.main.adapter.TidesRecyclerAdapter
+import com.mxmariner.main.model.TidesViewState
+import com.mxmariner.main.model.TidesViewStateLoadingComplete
+import com.mxmariner.main.model.TidesViewStateLoadingStarted
 import com.mxmariner.tides.repository.HarmonicsRepo
 import com.mxmariner.tides.ui.SnackbarController
 import com.mxmariner.tides.util.LocationPermissionResult
