+++ app/src/main/res/layout/activity_search_layout.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layout_above="@+id/navigation"
-    tools:context=".activity.LocationSearchActivity">
+    tools:context="com.mxmariner.main.activity.LocationSearchActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/appBarLayout"
