+++ app/src/main/res/layout/layout_tides_fragment.xml
