+++ app/src/main/res/layout/tide_list_view_holder.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.mxmariner.tides.main.view.StationListView
+<com.mxmariner.main.view.StationListView
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/stationListView"
     android:layout_width="match_parent"
