--- main/build.gradle
@@ -1,21 +1,0 @@
-apply plugin: "com.android.dynamic-feature"
-apply plugin: "kotlin-android"
-apply plugin: "kotlin-android-extensions"
-apply plugin: "kotlin-kapt"
-
-android {
-    compileSdkVersion sdk_version
-    defaultConfig {
-        minSdkVersion min_sdk_version
-        targetSdkVersion sdk_version
-        versionCode version_code
-        versionName version_name
-    }
-    kotlinOptions {
-        jvmTarget = JavaVersion.VERSION_1_8
-    }
-}
-
-dependencies {
-    implementation project(':app')
-}
