--- main/src/main/AndroidManifest.xml
@@ -1,44 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:dist="http://schemas.android.com/apk/distribution"
-        package="com.mxmariner.tides.main">
-
-    <dist:module
-            dist:instant="true"
-            dist:onDemand="false"
-            dist:title="@string/title_main">
-        <dist:fusing dist:include="false" />
-    </dist:module>
-
-    <application>
-        <activity
-                android:name=".activity.MainActivity"
-                android:theme="@style/AppTheme">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-            <intent-filter
-                    android:autoVerify="true"
-                    android:order="1">
-                <action android:name="android.intent.action.VIEW" />
-
-                <category android:name="android.intent.category.BROWSABLE" />
-                <category android:name="android.intent.category.DEFAULT" />
-
-                <data android:scheme="http" />
-                <data android:scheme="https" />
-                <data android:host="mxmariner.com" />
-                <data android:path="/tides" />
-                <data android:pathPattern="/tides?.*" />
-            </intent-filter>
-            <meta-data
-                    android:name="default-url"
-                    android:value="https://mxmariner.com/tides" />
-        </activity>
-        <activity
-                android:name=".activity.LocationSearchActivity"
-                android:theme="@style/TransparentTheme" />
-    </application>
-</manifest>
-
