--- main/src/main/res/values/strings.xml
@@ -1,3 +1,0 @@
-<resources>
-    <string name="title_main">MX Tides</string>
-</resources>
