+++ settings.gradle
@@ -1,4 +1,3 @@
 include ":app",
-        ":main",
         ":station",
         ":mxtide"
